requirements.txt
Pillow==9.0.1
discord.py==2.0.0
wargaming==2018.5.0
protobuf==3.18.3
pandas~=1.3.4
numpy~=1.21
pymongo~=4.0.1
dnspython==2.1.0
deprecation~=2.1.0
tqdm~=4.64.1
protobuf>=3.18.3
inflect~=6.0.2